First working version of java codegen daml interface support
