+++ language-support/java/codegen/src/main/scala/com/digitalasset/daml/lf/codegen/CodeGenRunner.scala
@@ -10,7 +10,7 @@ import com.daml.lf.archive.DarParser
 import com.daml.lf.codegen.backend.Backend
 import com.daml.lf.codegen.backend.java.JavaBackend
 import com.daml.lf.codegen.conf.{Conf, PackageReference}
-import com.daml.lf.data.{ImmArray, Ref}
+import com.daml.lf.data.ImmArray
 import com.daml.lf.data.Ref.PackageId
 import com.daml.lf.iface.reader.{Errors, InterfaceReader}
 import com.daml.lf.iface.{Type => _, _}
@@ -54,14 +54,14 @@ object CodeGenRunner extends StrictLogging {
     )
     val ec: ExecutionContext = ExecutionContext.fromExecutor(executor)
 
-    val (interfaces, astInterfaces, pkgPrefixes) = collectDamlLfInterfaces(conf)
-    generateCode(interfaces, astInterfaces, conf, pkgPrefixes)(ec)
+    val (interfaces, pkgPrefixes) = collectDamlLfInterfaces(conf)
+    generateCode(interfaces, conf, pkgPrefixes)(ec)
     val _ = executor.shutdownNow()
   }
 
   private[codegen] def collectDamlLfInterfaces(
       conf: Conf
-  ): (Seq[Interface], Map[Ref.TypeConName, DefInterface.FWT], Map[PackageId, String]) = {
+  ): (Seq[Interface], Map[PackageId, String]) = {
     val interfacesAndPrefixes = conf.darFiles.toList.flatMap { case (path, pkgPrefix) =>
       val file = path.toFile
       // Explicitly calling `get` to bubble up any exception when reading the dar
@@ -91,8 +91,7 @@ object CodeGenRunner extends StrictLogging {
           failIfUnresolvedChoicesLeft = true,
         )
       )
-    environmentInterface.astInterfaces.foreach(it => new Exception(it.toString()))
-    (fullyResolvedInterfaces, environmentInterface.astInterfaces, prefixes)
+    (fullyResolvedInterfaces, prefixes)
   }
 
   private[CodeGenRunner] def generateFile(
@@ -186,7 +185,6 @@ object CodeGenRunner extends StrictLogging {
 
   private[CodeGenRunner] def generateCode(
       interfaces: Seq[Interface],
-      astInterfaces: Map[Ref.TypeConName, DefInterface.FWT],
       conf: Conf,
       pkgPrefixes: Map[PackageId, String],
   )(implicit ec: ExecutionContext): Unit = {
@@ -207,11 +205,6 @@ object CodeGenRunner extends StrictLogging {
         _ <- Future.traverse(preprocessedInterfaceTrees.interfaceTrees)(
           processInterfaceTree(_, conf, prefixes)
         )
-        _ <- Future.successful {
-          astInterfaces.foreach { case (name, choices) =>
-            TemplateClass2.generate
-          }
-        }
 
       } yield ()
     }
